// src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  artist?: string | null;
  'artist-link'?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Archive {
  id: string;
  'archive-description'?: string | null;
  'slumber-text'?: string | null;
  fragments?:
    | {
        title: string;
        art?: (string | null) | Media;
        id?: string | null;
      }[]
    | null;
  'crystal-chamber-art'?: (string | null) | Media;
  'gathering-radiance'?:
    | {
        title: string;
        description?: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        id?: string | null;
      }[]
    | null;
  /**
   * The ID of the video you want to display, the part after https://www.youtube.com/watch?v= {video id here}
   */
  'future-lore'?: string | null;
  unlocks?:
    | {
        name: string;
        goal?: string | null;
        id?: string | null;
      }[]
    | null;
  'vaelorans-description'?: string | null;
  'jasper-descriptions'?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  'jasper-message'?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}